[add] добавлено не хватающая табличка [fix] float --> double в некоторых модельках

+++ AccountingScholarships.Domain/DTO/EpvoSso/EpvoSpecialitiesDto.cs
@@ -2,17 +2,17 @@ namespace AccountingScholarships.Domain.DTO.EpvoSso;
 
 public class EpvoSpecialitiesDto
 {
-    public float? UniversityId { get; set; }
-    public float? Id { get; set; }
-    public float? ProfCafId { get; set; }
+    public double? UniversityId { get; set; }
+    public double? Id { get; set; }
+    public double? ProfCafId { get; set; }
     public string? NameRu { get; set; }
     public string? NameKz { get; set; }
     public string? NameEn { get; set; }
     public bool? DoubleDiploma { get; set; }
     public bool? JointEp { get; set; }
     public string? SpecializationCode { get; set; }
-    public float? StatusEp { get; set; }
-    public float? EduProgType { get; set; }
+    public double? StatusEp { get; set; }
+    public double? EduProgType { get; set; }
     public bool? Default { get; set; }
     public bool? Interdisciplinary { get; set; }
     public bool? EducationProgram { get; set; }
